+++ src/data_source/storage/fs.rs
@@ -17,6 +17,8 @@ use super::{
     pruning::{PruneStorage, PrunedHeightStorage, PrunerConfig},
     AvailabilityStorage,
 };
+use crate::merklized_state::{MerklizedStateDataSource, Snapshot};
+use crate::QueryError;
 use crate::{
     availability::{
         data_source::{BlockId, LeafId, UpdateAvailabilityData},
@@ -30,10 +32,13 @@ use crate::{
     types::HeightIndexed,
     Header, MissingSnafu, NotFoundSnafu, Payload, QueryResult, VidCommitment, VidShare,
 };
+use ark_serialize::CanonicalDeserialize;
 use async_trait::async_trait;
 use atomic_store::{AtomicStore, AtomicStoreLoader, PersistenceError};
 use commit::Committable;
 use hotshot_types::traits::{block_contents::BlockHeader, node_implementation::NodeType};
+use jf_primitives::merkle_tree::{prelude::MerklePath, Element, Index, NodeValue, ToTraversalPath};
+
 use serde::{de::DeserializeOwned, Serialize};
 use snafu::OptionExt;
 use std::collections::{
@@ -44,6 +49,7 @@ use std::convert::Infallible;
 use std::hash::Hash;
 use std::ops::{Bound, RangeBounds};
 use std::path::Path;
+use typenum::Unsigned;
 
 const CACHED_LEAVES_COUNT: usize = 100;
 const CACHED_BLOCKS_COUNT: usize = 100;
@@ -569,5 +575,29 @@ where
         }
 
         Ok(res)
+    }
+}
+
+#[async_trait]
+impl<Types: NodeType> MerklizedStateDataSource<Types> for FileSystemStorage<Types>
+where
+    Payload<Types>: QueryablePayload,
+{
+    type Error = QueryError;
+
+    async fn get_path<
+        E: Element + Send + DeserializeOwned,
+        I: Index + Send + ToTraversalPath<A> + DeserializeOwned,
+        A: Unsigned,
+        T: NodeValue + Send + CanonicalDeserialize,
+    >(
+        &self,
+        _state_type: &'static str,
+        _tree_height: usize,
+        _header_state_commitment_field: &'static str,
+        _snapshot: Snapshot<Types>,
+        _key: serde_json::Value,
+    ) -> QueryResult<MerklePath<E, I, T>> {
+        Err(QueryError::NotFound)
     }
 }
