+++ src/merklized_state/data_source.rs
@@ -11,9 +11,8 @@
 // see <https://www.gnu.org/licenses/>.
 
 use crate::{
-    availability::QueryablePayload,
-    data_source::storage::{sql::*, FileSystemStorage, NoStorage, SqlStorage},
-    Payload, QueryError, QueryResult,
+    data_source::storage::{sql::*, SqlStorage},
+    QueryError, QueryResult,
 };
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize, SerializationError};
 use async_std::stream::StreamExt;
@@ -410,51 +409,6 @@ impl<Types: NodeType> MerklizedStateDataSource<Types> for SqlStorage {
         }
 
         Ok(proof_path)
-    }
-}
-
-#[async_trait]
-impl<Types: NodeType> MerklizedStateDataSource<Types> for FileSystemStorage<Types>
-where
-    Payload<Types>: QueryablePayload,
-{
-    type Error = QueryError;
-
-    async fn get_path<
-        E: Element + Send + DeserializeOwned,
-        I: Index + Send + ToTraversalPath<A> + DeserializeOwned,
-        A: Unsigned,
-        T: NodeValue + Send + CanonicalDeserialize,
-    >(
-        &self,
-        _state_type: &'static str,
-        _tree_height: usize,
-        _header_state_commitment_field: &'static str,
-        _snapshot: Snapshot<Types>,
-        _key: Value,
-    ) -> QueryResult<MerklePath<E, I, T>> {
-        Err(QueryError::NotFound)
-    }
-}
-
-#[async_trait]
-impl<Types: NodeType> MerklizedStateDataSource<Types> for NoStorage {
-    type Error = QueryError;
-
-    async fn get_path<
-        E: Element + Send + DeserializeOwned,
-        I: Index + Send + ToTraversalPath<A> + DeserializeOwned,
-        A: Unsigned,
-        T: NodeValue + Send + CanonicalDeserialize,
-    >(
-        &self,
-        _state_type: &'static str,
-        _tree_height: usize,
-        _header_state_commitment_field: &'static str,
-        _snapshot: Snapshot<Types>,
-        _key: Value,
-    ) -> QueryResult<MerklePath<E, I, T>> {
-        Err(QueryError::NotFound)
     }
 }
